[!!!][TASK] Simplify PackageManagement / Remove Flow
The PackageManager was originally backported from Flow
but includes too many classes and options that the Core does not
use or cannot use at all.

The Package code is thus merged with only the parts needed
for the TYPO3 Core, and adapted accordingly. All Flow classes
are either removed or merged into the needed Package classes
of the TYPO3 Core.

In order to speed up the bootstrap and prepare for tight composer
integration, this task needs to be sorted out first.

Next step is to further cleanup code within the Package*
classes themselves - things that are not called at all in
the TYPO3 Core.

Resolves: #67027
Releases: master
Change-Id: Ia6823374afa515902e3d5237cef596a440747e3b
Reviewed-on: http://review.typo3.org/39518
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>