[BUGFIX] Prevent wrong record copies
Before starting any record copy process, we have to check if the
destination page can handle the record type. Currently if you copy a
record with IRRE references, the references are copied first but the
original record isn't copied because of missing permissions. The
copied IRRE elements stay and pollute the original record as the
references are added to it. This patch adds a check if the destination
page can handle the mother record type and starts copy process only if
allowed.

Resolves: #66920
Releases: master, 6.2
Change-Id: I274792b13170f37e3e799ff892ea877ba44d90a8
Reviewed-on: http://review.typo3.org/39616
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>