[BUGFIX] Check page access only if integer in ModuleRunner
If you register a module in main module "file" an error is thrown
because the ModuleRunner checks page access for the variable "id".
In any file module the id is a FAL identifier which can't be converted
to an integer value and the check fails.

This patch adds a limitation to the ModuleRunner to check page access
only if the id can be interpreted as integer value.

Releases: master, 6.2
Resolves: #67079
Change-Id: Iba44499b9b13172818aee48aefb01d102f810285
Reviewed-on: http://review.typo3.org/39648
Reviewed-by: Markus Sommer <markussom@posteo.de>
Tested-by: Markus Sommer <markussom@posteo.de>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>