[BUGFIX] Adhere "suggestions" when resolving ext loading order
The issue #58701 introduced the functionality to fetch suggestions defined
in composer.json or ext_emconf.php into the Package class.
A copy and paste error causes the loading of the suggestions to
actually fail, so those never reach the DependencyResolver.

This patch fixes the tiny mistake and thus ensures that suggestions
really influence the loading order of extensions.

Resolves: #51069
Releases: master, 6.2
Change-Id: Ib12e7f91f49c0b2bda56fdb4394a604e322b741c
Reviewed-on: http://review.typo3.org/39711
Reviewed-by: Xavier Perseguers <xavier@typo3.org>
Tested-by: Xavier Perseguers <xavier@typo3.org>