[TASK] Do not generate new processed file if it already exists
When a processed file already exists but the sys_file_processedfile
record is deleted the whole processing task is executed again even
if the file already exists.

This change first checks if `targetFile` already exists and if so
uses that file and marks the task as done.

Currently this only works for `Local` file storages as the FAL API
misses proper methods get a file by folder identifier and file name.
When https://review.typo3.org/38479/ is merged a follow up is
needed for master #67126.

Resolves: #67095
Releases: master
Change-Id: I352428ddaf1d2afe4bbf031992b484ade3635e4a
Reviewed-on: http://review.typo3.org/39709
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Frank Nägler <typo3@naegler.net>
Tested-by: Frank Nägler <typo3@naegler.net>