[FOLLOWUP][BUGFIX] Show full folder path in file list title
In old situation the full path to of the folder was shown.
With #64759 only the accessible part is shown. The logic
to determine what the root is of the 'current' filemount
is not suitable for this and will always return the root path of
the first filemount within the storage. This could result in a wrong
root identifier.

This patch fixes the output of the readable path to the
full path by using the real storage root instead of the storage root
defined by the filemount.

Resolves: #67148
Related: #64759
Releases: master, 6.2
Change-Id: Ie00fa2dc7d3833f798c5a2913bb9674b2b15b6e7
Reviewed-on: http://review.typo3.org/39787
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>
Reviewed-by: Xavier Perseguers <xavier@typo3.org>
Tested-by: Xavier Perseguers <xavier@typo3.org>