[BUGFIX] Prevent information disclosure in file list
Currently the doc header title of a folder shows the full path even
if the current folder is inside a mount point. This patch prevents
showing the full path by not disabling the permission check but catch
a thrown exception and return the path inside the mount point.

Resolves: #67245
Releases: master, 6.2
Change-Id: I6e5486e8c6f923decc4016b57ff60a562f189749
Reviewed-on: http://review.typo3.org/39898
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Harry Glatz <glatz@analog.de>
Reviewed-by: Susanne Moog <typo3@susannemoog.de>
Tested-by: Susanne Moog <typo3@susannemoog.de>