[BUGFIX] Fix ambiguous identifier for imagesizes cache
Use sha1 of filepath instead of filemtime and filesize as identifier for
cached image dimensions to enable invalidation of stale cached entries
when an image is altered.
This also fixes cache collisions when dimensions are cached for a large
number of images with equal mtimes, some of which may also have equal
filesize, but different dimensions.

Releases: master
Resolves: #67289
Related: #28484
Change-Id: I520a8292dfbf3beb8efae996d286471e641ffd07
Reviewed-on: http://review.typo3.org/39979
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>