[TASK] Nail down php 5.6 instantiation issues in tests
We know PHP 5.6 has issues in 6.2 in extbase object creation.
The patch adds the unit test from the "instantiator" patch
of issue #66473. This test fails, so travis-ci is configured
with PHP 5.6 as known failure at the moment.

With the instantiator patch applied, initializeObject() was
no longer called in some PHP versions, another test is added
to verify this regression can not show up again.

The test bed currently ensures:
* PHP 5.6 fails with current version
* PHP 5.4, 5.5 and 5.6 would fail if only "instantiator" patch
from #66473 is applied
* PHP 5.3 fails with "initializeObject is called twice" if patch
from #67274 is applied

Later is to get 5.6 green again with no further failure in other
PHP versions with other patches in 6.2 without changing tests.

Resolves: #67408
Related: #66473
Related: #67274
Releases: 6.2
Change-Id: Id4e81877e696dd1abc5b39497a19fa20a702048c
Reviewed-on: http://review.typo3.org/40163
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>