[BUGFIX] Always use MCRYPT_DEV_URANDOM if using mcrypt
Using MCRYPT_RAND was introduced because of a bug in PHP versions lower
than 5.3.7 on Windows in combination with IIS.

Since we require higher PHP versions in all maintained versions we can
remove this workaround and use MCRYPT_DEV_URANDOM again. By doing so we
fix a bug on Windows caused by not enough randomness.

Releases: 6.2, master
Resolves: #53034
Change-Id: Ibe74eb0277934e9300ffd9b00cc89a5f8bb008fb
Reviewed-on: http://review.typo3.org/40258
Reviewed-by: Stephan Großberndt <stephan@grossberndt.de>
Tested-by: Stephan Großberndt <stephan@grossberndt.de>
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>