[BUGFIX] FormEngine: Size equals to 1 if none set on multi-selects
Having a FormEngine select element with maxitems=10
and no size set, then a typical single-select element
is rendered, with no chance of using the up- and
down buttons provided, as the single select does not
allow to use more than one entry in rendering.

The browser behavior is not adopted and if size
is not given, a broken state is rendered.

Releases: master
Resolves: #56259
Change-Id: I9963503cb046353b81f13b0c624bddbd7e5b74f8
Reviewed-on: http://review.typo3.org/27829
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>