[BUGFIX] Include doktype in pageOverlayFields
Translation overlays must include the doktype field of the translated
page, to take into account page type changes. Otherwise we get unwanted
behaviour if e.g. the original language is a shortcut, but the
translated page is a standard page. Then still a shortcut will be
rendered, with weird possible mode combinations, rendering the page
unviewable.

This is especially nasty as the translation mode is reset when
translating, making the translated page a "shortcut to selected page"
which is always broken (!) if no page is selected.

Releases: master
Resolves: #67562
Change-Id: I8b562643de6c0fd0e267875e298f7c6de14ebbe3
Reviewed-on: http://review.typo3.org/40365
Reviewed-by: Daniel Maier <dani-maier@gmx.de>
Tested-by: Daniel Maier <dani-maier@gmx.de>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>