[CLEANUP] Code duplication in DataHandler::copyRecord_processManyToMany()
There was a foreach loop that only differed in one line, but was
duplicated due to the complex condition around it. This commit splits
the condition, introduces helper variables and moves the condition
inside the foreach loop to get rid of the duplication.

Additionally, duplicated lines for purging the items of the relation
analyzer are also removed.

Resolves: #67579
Releases: master
Change-Id: Ibb9277dfb32960cc77aee17839d7df3bf57fc6da
Reviewed-on: http://review.typo3.org/40423
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Stephan Großberndt <stephan@grossberndt.de>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>