[!!!][TASK] Remove PHP library inclusion in FE
The functionality to add PHP include files
during the Frontend request is quite old and
obsolete as there are plenty of hooks available.

The functionality is in this case obsolete as it does
not apply to the design patterns we apply nowadays.

This patch removes the the library inclusion and
the associated TypoScript options and PHP function.

Resolves: #67646
Releases: master
Change-Id: I376d0ed5eed324a1a0202021529ed526a48edab4
Reviewed-on: http://review.typo3.org/40484
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Daniel Maier <dani-maier@gmx.de>
Tested-by: Daniel Maier <dani-maier@gmx.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>