[FEATURE] Introduce Data Processor for Files
In order to use real file objects within Fluid templates or other
places with cObjects, a new data processor for Files is added.

As it does most of the same part (collecting files from various
sources) as the FILES cObject, the logic has been separated
into an own class, the so-called "File Collector".

Resolves: #67662
Releases: master
Change-Id: I05c494db65caa7efd1847504a057b3b21efb8702
Reviewed-on: http://review.typo3.org/40508
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>