[BUGFIX] Increase lifetime of cache_rootline
The default lifetime of frontend related rootline cache is
only one hour, raising query load in production. This is
raised to 30 days now - same as cache_pagesection.
The rootline cache entries are tagged with pageId_* and
in cache group "pages", so the DataHandler automatically
drops those entries if pages are moved around or change.

Resolves: #50231
Releases: master, 6.2
Change-Id: I3839982058f8b78726c050d5a7da90c0f2588102
Reviewed-on: http://review.typo3.org/40666
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>