[TASK] Include autoload info of all core extensions
Previously classes for extensions that were not loaded,
were also not autoloadable. To simplify composer installation
of the core, we now include autoload information of all extensions
that are delivered with the typo3/cms package.

By doing so, authors of distribution composer.json files
do not need to amend it with class loading information of
core extensions any more.

This means that class loading information for every
system extension will be written to typo3/contrib/vendor
by the packaging script, so that we do not need to respect
these when building additional autoload information for extensions
that are installed via the extension manager in a non composer
managed environment.

Releases: master, 7.3
Resolves: #67767
Change-Id: I1b68ebff29fc5392a0d874a17b9caa3dd4483096
Reviewed-on: http://review.typo3.org/40674
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Reviewed-by: Daniel Goerz <ervaude@web.de>
Tested-by: Daniel Goerz <ervaude@web.de>
Reviewed-by: Sascha Egerer <sascha@sascha-egerer.de>
Tested-by: Sascha Egerer <sascha@sascha-egerer.de>