[TASK] Include autoload info of all core extensions
Previously classes for extensions that were not loaded,
were also not autoloadable. To simplify composer installation
of the core, we now include autoload information of all extensions
that are delivered with the typo3/cms package.

By doing so, authors of distribution composer.json files
do not need to amend it with class loading information of
core extensions any more.

This means that class loading information for every
system extension will be written to typo3/contrib/vendor
by the packaging script, so that we do not need to respect
these when building additional autoload information for extensions
that are installed via the extension manager in a non composer
managed environment.

Releases: master, 7.3
Resolves: #67767
Change-Id: I1b68ebff29fc5392a0d874a17b9caa3dd4483096
(cherry picked from commit 023a30bfdda5573f2acaf7d8826ae5b1b063ae6c)
Reviewed-on: http://review.typo3.org/40676
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>