[BUGFIX] Only count items in ForViewHelper when ``iteration`` argument is set
When using the ``iteration`` argument, the ``ForViewHelper`` provides
information about the current iteration, including the ``cycle`` and
``total``. Therefore it needs to count all elements.
The problem is, that this happens even if the iteration argument is
not specified leading to performance issues especially when iterating
over a (subset of a) lot of (remote) items.

This patch fixes this by adding a check for the ``iteration`` argument.

Resolves: #67801
Releases: master, 6.2
Change-Id: I4645ed03322bea36f2c36263378900f401626f24
Reviewed-on: http://review.typo3.org/40718
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>