[BUGFIX] Clean up and fix class loading
With class loading information added, we must also add class alias information.
Besides that, we need to distinguish between extensions that should be
loadable and are part of the framework and those which should only be
loaded optionally and are subject to be removed from the core in the near future.

These extensions (namely compatibilty6 and mediace) are now made "regular"
TYPO3 extensions and class loading information is removed from the main composer.json

We now also add the class alias map information to the sub-packages delivered
by the core to have a clean state once they are extracted by a subtree split.

Having all this in place, we can unify the check for framework extensions
in the dependency resolver as well.

Releases: 7.3, master
Resolves: #67933
Change-Id: I08b576c96921058afcf5e03a7c5c3aa6e25d2d55
Reviewed-on: http://review.typo3.org/40961
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>