[TASK] Remove NEWS.md
NEWS.md is just a dummy file pointing to the Changelog .rst
information. There is not much benefit in it, so it is removed
along with the reference from README.md now.

Since there is no stable link to the wiki pages with non
volatile information, it also does not make much sense to
link to them. This may however change in the future and
a hint to a list of newest changes within README.md, a file
that is for instance displayed by default on github, would
be appreciated.

Resolves: #68032
Releases: master
Change-Id: I1a6d91e8d716b193bf81df20a28541c3a4da1faf
Reviewed-on: http://review.typo3.org/41067
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>