[BUGFIX] Ensure PageRenderer singleton is set
Before rendering _INT content objects, the page renderer is unserialized
and stored in a TypoScriptFrontendController property. However this new
object is not updated in the singleton store in GeneralUtility.

This leads to the situation that plugin code that (legally) just instantiate
the page renderer using GeneralUtility::makeInstance() receive the wrong object.

Make sure the singleton store is updated once the page renderer is unserialized
for _INT processing.

Releases: master, 6.2
Resolves: #68061
Change-Id: I162b8a4dc102ffd8cc0fcc4d21ccc271c861b187
Reviewed-on: http://review.typo3.org/41112
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>