[BUGFIX] Improve detection of installed setup
When installing TYPO3 via composer and installing
an extension at the same time, the install tool cannot
be called after that, as the typo3conf/ directory already
exists.

The check is optimized to check for the LocalConfiguration.php
and PackageStates.php which are not present at first install.

Resolves: #68261
Releases: master
Change-Id: Ib2a787bf0968b1bcbcc3ed02e0b678ff8a74aa91
Reviewed-on: http://review.typo3.org/41518
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Jan Helke <typo3@helke.de>
Reviewed-by: Susanne Moog <typo3@susannemoog.de>
Tested-by: Susanne Moog <typo3@susannemoog.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>