[BUGFIX] FormEngine warning
The patch fixes a todo that was marked as code smell already.
The FormEngine class construct relies on a defined return structure
for child nodes. This must not be violated and may throw all kind
of warnings.
An edge case is fixed, where the InlineRecordContainer is called
by FormEngine ajax inline "extend" or by the InlineControlContainer,
those parents must react differently, so this communcation is now
done with an exception.

Resolves: #68262
Resolves: #67592
Releases: master
Change-Id: I32996e31d27be3d0556b9b68c7291624e0d289d2
Reviewed-on: http://review.typo3.org/41521
Reviewed-by: Susanne Moog <typo3@susannemoog.de>
Tested-by: Susanne Moog <typo3@susannemoog.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>