[BUGFIX] Reuse local mirror.xml.gz when updating translations
When updating translation packages, the local mirrors.xml.gz should be
reused instead of being fetched again and again.

The list of mirrors virtually never changes and is already updated when
fetching the list of TER extensions in Extension Manager. It should thus
be assumed that the local content is just correct and may be used safely.

Change-Id: I5863762061b2a55d2439a92d53fdbf02bcd6e124
Resolves: #68364
Releases: master, 6.2
Reviewed-on: http://review.typo3.org/41722
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Susanne Moog <typo3@susannemoog.de>
Tested-by: Susanne Moog <typo3@susannemoog.de>