[FOLLOWUP][BUGFIX] List module must not do editlock for admins
Commit 3e2d72d3 introduced a generic editlock mechanism for all record
types. If a page is locked for editing, all the content elements are
also locked.

The list module related part of that commit introduced a bug, which is
fixed with this commit: the lock was also applied for admins, hence the
edit icons were removed also for them.

Change-Id: I801544a6b3f571e10c8e41366c073b352bc22c8f
Releases: master
Related: #68279
Reviewed-on: http://review.typo3.org/41738
Reviewed-by: Susanne Moog <typo3@susannemoog.de>
Tested-by: Susanne Moog <typo3@susannemoog.de>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>