[!!!][TASK] Append hash to default log file name
A hash is appended to the default log file name the FileWriter uses
when no other log file is provided in the configuration.

To make this work the logFile configuration is removed for the
writerConfiguration section.

Additionally the "deprecated" log configuration section is removed
because it is currently not used by the core.

Resolves: #52705
Releases: master
Change-Id: I1a55ea2b9eb81e1298dea22dd62ec575bd3f4dcc
Reviewed-on: http://review.typo3.org/41633
Reviewed-by: Steffen Müller <typo3@t3node.com>
Tested-by: Steffen Müller <typo3@t3node.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>