[BUGFIX] Ensure jQuery is loaded as well when loading the Lang module
If we forcefully load the JS Lang-module, which depends on jQuery, we
have to ensure that jQuery is requested as well, otherwise the dependency
of the Lang-module fails.

Note that jQuery is never loaded via requireJS. jQuery must be explicitly added
by calling  the loadJquery method of PageRenderer. jQuery itself registers as
an AMD module and can then be used as dependency by other AMD modules.

We therefore also remove the (completely useless) path definition in the
requireJS configuration.

Releases: master
Resolves: #68555
Change-Id: If8ef5b1b79523cf00451f3ba34a00b983945c000
Reviewed-on: http://review.typo3.org/41981
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Tomas Norre Mikkelsen <tomasnorre@gmail.com>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Tested-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>