[BUGFIX] Scheduler: make lastexecution_failure field non-mandatory
MySQL in strict mode requires TEXT fields defined with the NOT NULL
attribute to be explicitly set on INSERT statements. Due to execution
failures of scheduler jobs being the exception and not the rule as well
as MySQL not allowing DEFAULT values for TEXT colums changing the field
to not be mandatory is sensible.

Resolves: #68418
Releases: master
Change-Id: I0ed3a3b70d4f61f8e418741ae76aa0010b6c68a3
Reviewed-on: http://review.typo3.org/41994
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Tomas Norre Mikkelsen <tomasnorre@gmail.com>
Tested-by: Tomas Norre Mikkelsen <tomasnorre@gmail.com>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>