[BUGFIX] Logging: Make sys_log NEWid column larger
Update the table definition for the sys_log table to store the whole
NEWid value from TCEMAIN. NEWid typically is 'NEW' with a uniqid of 23
characters appended, so 20 characters is not enough to store the whole
id. For MySQL not running in strict mode this fixes silent truncation
of the data, with strict mode it fixes a "data too long for column"
error.

Resolves: #68564
Releases: master
Change-Id: Id89bc9c5759974379da20f6af8b945e5f4fbefc6
Reviewed-on: http://review.typo3.org/41996
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Tomas Norre Mikkelsen <tomasnorre@gmail.com>
Tested-by: Tomas Norre Mikkelsen <tomasnorre@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>