[BUGFIX] mergeIfNotBlank does not work for recordTitle and tcaType
If a field is TCA config is set to l10n_mode=mergeIfNotBlank, the
value from the default language record is used as fallback if the
overlay value is blank.
Apart from the fact that "mergeIfNotBlank" is very misleading
wording, this is not handled correctly at two places in the backend.

Resolves: #68592
Releases: master
Change-Id: If112cbb5cc534a4c2d3379832c37e6bbe3def4d8
Reviewed-on: http://review.typo3.org/42050
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>