[BUGFIX] displayCond in Flexform sheets
This actually fixes two independent bugs:

1) The condition matcher could not access information from another
sheet, because only the current sheet’s data was handed over. That is
simply wrong, as no information depending on a sheet name should be used
in the condition (otherwise you could hide the sheet you’re seeing and
never be able to get it back).

2) The field name in the split condition was changed, but the changed
condition was not passed to the condition matcher.

Resolves: #67879
Releases: master
Change-Id: Iecc194dddc749d70cbe13b52fa4aff62c63ace3c
Reviewed-on: http://review.typo3.org/42071
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>