[BUGFIX] Remove wrong references in Install Tool ViewHelper
The ContainerViewHelper from the install tool has the option
to enable the click menu, which references to a file which is
now an AMD module (= needs to be loaded via RequireJS).

Additionally, the ExtJS logic is removed from the install tool
completely as well, as it is unused.

Releases: master
Resolves: #68628
Change-Id: I1aab35c49e98bd2fb42b16f5fe4a16c7b28acb92
Reviewed-on: http://review.typo3.org/42097
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>