[TASK] Disallow access to vcs directories
Access to vcs directories should be denied by default to avoid leaking
information to the outsite that might have been accidentally committed
to a repository.
This does not effect the core as the version information is public by
license, but any repository containing configuration or third party
extensions.

Resolves: #68626
Releases: master,6.2
Change-Id: I4f3cb88e577f56ac71d882c8b11341da46a8b757
Reviewed-on: http://review.typo3.org/42100
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>