[BUGFIX] Rename setLocale() in DateTimeViewHelperTest
The method setLocale($locale) in the DateTimeViewHelperTest
was implemented to test against several locales. However,
as there is a setLocale() in the parent class from PHPUnit that
does not take any parameter, PHP7 throws a warning as the
subclass does not have the same method signature as the parent
class.

The patch changes the method to be named differently thus,
solving the method signature conflict inside the test.

Resolves: #67092
Releases: 6.2, master
Change-Id: Ia05d90c64a3a8e2fc599935d2435192392e5a667
Reviewed-on: http://review.typo3.org/42113
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>