[TASK] Rename ResourceStorage::processUploadedFilename to sanitizeFileName
Feature #67545 introduced ResourceStorage::processUploadedFilename and
changed the existing PreFileAddSignal. The name of the newly introduced
method doesn't really fit its functionality and using the same signal
for
two different purposes is also not correct.

This change renames the method
processUploadedFilename() to sanitizeFileName()
and introduces the SanitizeFileNameSignal.
The changes made to PreFileAddSignal are reverted.

Resolves: #68600
Related: #67545
Releases: master
Change-Id: Iabeffe5b8b3fe85a4f49b9341a45b2a026dd241c
Reviewed-on: http://review.typo3.org/42116
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>