[BUGFIX] PSR-7: Do not let legacy code take precedence over HTTP headers
The code to not overwrite regular headers set by PSR-7 response
is obsolete and should be removed. Legacy code does not have access
to the response object, thus the PSR-7 header list would always be empty.

We also do not want to give legacy code precedence here if our code
explicitly sets some headers.
Also make sure that, if legacy code added some headers, which are also
set in the response object (by us), that the headers are not additionally added
but override the legacy headers.

The only exception is the status code header, which we must allow to
be changed from legacy code for the time being.

Resolves: #68670
Related: #68506
Releases: master
Change-Id: I792f18debcadebc5d9b686a8df2de5a1f547c8ad
Reviewed-on: http://review.typo3.org/42161
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>