[BUGFIX] DataHandler: cast value to string in checkValueForInput()
Explicitly cast the value as a string before using substr() on it.
Otherwise NULL values would be passed to substr(), resulting in a return
value of FALSE. Without MySQL strict mode this has no direct influence
on the further usage of the value as string as the string representation
of FALSE is an empty string. With MySQL strict mode enabled this leads
to errors upon insertion of records into the database as boolean values
get converted to 0/1 respectively.

Resolves: #68602
Related: #68562
Related: #55082
Releases: master
Change-Id: Idc3ffa147965083492a04b12c64a6c15ddd2277b
Reviewed-on: http://review.typo3.org/42260
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Tested-by: Frank Nägler <frank.naegler@typo3.org>