[BUGFIX] Handle localizeReferencesAtParentLocalization for group/db
TCA config property 'localizeReferencesAtParentLocalization'
for database relations is only handled with type=select
having a 'foreign_table' defined.

With the similar type=group, this is only handled if
'foreign_table' is set, which is odd since group is based
on 'allowed' over 'foreign_table'.

The change adapts DataHandler to consider 'allowed' with
type=group tables and checks whether at least one of them
is localizable.

This solves an @todo from #94570 in the group DataHandling
tests. The test set is completed by adding a similar set
for type=select.

Resolves: #41713
Related: #94570
Releases: master
Change-Id: I33df58fd4c9bdf31e27f64532c371c74b3f52627
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72102
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>