[BUGFIX] Shortcut list (top right) is filled by duplicated entries
The AjaxShortcut function has no validation of an already existing entry in the
shortcut list before it create a new one. At this case you have multiple
duplicated entries in your shortcut list and that is simply not expected and
makes no sense.

Additional behaviour for "Delete" buttons in the shortcut list:
After clicking a "Delete" action of a shortcut list entry you get a
modal window to confirm. But the gui blocker for the modal window cannot
completely disallow the shortcut buttons outside the modal context. In this case
the zIndex for the "modal" class (with gui blocker) is simply wrong here
(too small).

Resolves: #68297
Releases: master
Change-Id: I3d10e25a90055d935e33c005274de87d6de098ef
Reviewed-on: http://review.typo3.org/42279
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Tested-by: Frank Nägler <frank.naegler@typo3.org>