[FEATURE] Show only distributions for current version
As the list of distributions gets longer, it is filtered now to
those distributions that suite the current TYPO3 version.

Distributions that do not comply with version constraints are not
shown anymore. A button can be pressed to get the full list again,
for people who know what they’re doing.

Resolves: #68724
Releases: master
Change-Id: I11c6846e706ba18c130677c65ec568f80e502ed8
Reviewed-on: http://review.typo3.org/42283
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Sascha Wilking <sascha.wilking@hmmh.de>
Tested-by: Sascha Wilking <sascha.wilking@hmmh.de>
Reviewed-by: Andreas Wolf <andreas.wolf@typo3.org>
Tested-by: Andreas Wolf <andreas.wolf@typo3.org>