[TASK] Tests: Allow NULL for optional fields in test extensions
The extensions irre_tutorial and test_datahandler extend the tt_content
table with fields which are not used by all tt_content elements. Due to
the fields being defined as NOT NULL but not having a default value this
leads to problems creating records when MySQL is running in strict mode.

This patch removes the NOT NULL restriction as BLOB/TEXT columns can not
have default values.

Resolves: #69000
Releases: master
Change-Id: Iefc7ae3b52e3b522c7ea70cfe84a927fd975fb6c
Reviewed-on: http://review.typo3.org/42623
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>