[BUGFIX] Tests: Create NEWid values the same way as DataHandler
The NEWid values generated in the functions tests of the core use a
different kind of NEWid value that is longer than the values created
within DataHandler. The length of the values generated in the tests
exceeds the length of the database column in the sys_log table, leading
to errors when running MySQL in strict mode.

This change uses the same code as DataHandler to generate NEWid values.

Resolves: #69039
Releases: master
Change-Id: I4380fb333a15d37b9a48b98b211a9143286ec73b
Reviewed-on: http://review.typo3.org/42625
Reviewed-by: Stephan Großberndt <stephan@grossberndt.de>
Reviewed-by: Alexander Opitz <opitz.alexander@googlemail.com>
Tested-by: Alexander Opitz <opitz.alexander@googlemail.com>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>