[BUGFIX] Fix dependencies for t3x upload
When extracting the t3x archive, the EM_CONF data is taken to
create a new ext_emconf.php file.
An already extracted ext_emconf.php file is overwritten in this process.

Since those t3x files are created by TER server and this TER server
does not include the extension constraints in this array,
all constraints are lost in this process.

To fix this, it is needed to use a present ext_emconf.php file
and merge the meta properties with its properties to construct
the final ext_emconf.php.

Resolves: #69070
Releases: master, 6.2
Change-Id: I3997bee05f256a3ff23f0eee516ecf0a323017c3
Reviewed-on: http://review.typo3.org/42642
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>