[BUGFIX] Tests: Fix Tag model config in extbase blog_example test extension
The Tag model in the extbase definition included a partial configuration
for translatable records. This includes NOT NULL fields in the SQL
schema definition that aren't declared in the TCA as well as fields that
are not being used by TYPO3 at all. This patch removes the partial
configuration as it caused errors when used with MySQL strict mode.

Resolves: #69089
Releases: master
Change-Id: Ib2c01ddfbbf4c7e4f9c0e53c324a3140ce3fa70f
Reviewed-on: http://review.typo3.org/42652
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>