[BUGFIX] Extbase: Optional arguments must not throw an Exception
Example:
public function barAction(\Vendor\Ext\Domain\Model\Foo $foo = NULL)

If the property mapper receives NULL, 0 or a valid value
the property mapping does not fail. But if the value is
a valid uid (>0) and there is no record with that uid
the whole property mapping fails.

While this is totally valid for required arguments,
optional arguments must be skipped.

Resolves: #55861
Resolves: #55837
Resolves: #62951
Releases: master, 6.2
Change-Id: I4c954bac50be1d34ea09bfc21f8978e9165683d1
Reviewed-on: http://review.typo3.org/42663
Reviewed-by: Alexander Schnitzler <typo3@alexanderschnitzler.de>
Tested-by: Alexander Schnitzler <typo3@alexanderschnitzler.de>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>