[TASK] Use 'wizard_element_browser' instead of 'browser' in FormEngine
FormEngine is the last place using the module 'browser'.
This module is the ElementBrowserFramesetController,
which uses a frameset. The frameset has no further functionality
and the additional JS provided by the class, is not used anymore.

We can therefore use the 'wizard_element_browser' module directly,
which does not render a frameset.

Resolves: #69140
Releases: master
Change-Id: Ibfc8147a6e822655ebd2e8f347f6f55c0e8382e5
Reviewed-on: http://review.typo3.org/42692
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Alexander Opitz <opitz.alexander@googlemail.com>
Tested-by: Alexander Opitz <opitz.alexander@googlemail.com>
Tested-by: Benni Mack <benni@typo3.org>