[BUGFIX] Show first tab in Element Wizard as fallback
Using the extension gridelements it is possible to restrict the
selectable content elements instead of the full menu configured
by TSconfig. At first, you might have 5 menus available and you
click on the last menu tab (maybe the fifth), so the backend
stores 5 as current tab.
Adding content elements within a gridelement, which restricts the
range of selectable CEs, you can only select the first tab (p.e.).
The original code assumes that the fifth tab is still there and
tries to select it.

This patch checks first against the existence of the selected tab.
If there is no "fifth" tab p.e. the index is set to 1.

Resolves: #59810
Releases: master, 6.2
Change-Id: I829489737de4631dfe3362e4fb11909be9315864
Reviewed-on: http://review.typo3.org/42799
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>