[TASK] Extract data overlay handling to common class
The data overlay handling of RelationHandler is extracted to
a separate common class. The scope of this object shall be
independent from frontend-only or backend-only constraints.

The new handler resolves version overlays for given IDs of
a database table and applies sorting for the target context.

This is just an intermediate solution to overcome sorting issues
of relations handled in RelationHandler. The goal of DataResolvers
should be to provide independent APIs for OneToMany and ManyToMany
relation resolving in the future as well.

Resolves: #69967
Releases: master, 6.2
Change-Id: I5a7f76a55b4af93aee6270168285fff9486a2e73
Reviewed-on: http://review.typo3.org/39901
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>