[!!!][TASK] Fix configuration for backend module icons
http://review.typo3.org/38004 removed the typo3/mod directory
with the main modules and their configuration and added the configuration
to `$GLOBALS['TBE_MODULES']`directly.
However for that a sub array key `configuration` was used, which is
inconsistent with all our current API and its usages across other
modules, leading to the situation that this configuration was basically
ignored.

On top of that, http://review.typo3.org/39316 introduced the possibility
to use sprite icons as module icons, which used the configuration key
`icon` in the above mentioned sub array key `configuration`,
making this key the only option evaluated in the sub array.

This change removes this sub array and moves the sprite icon
configuration on top for consistency. For that the key must be changed
to 'iconIdentifier' as 'icon' was already taken for Extbase module
icon configuration.

Resolves: #69291
Releases: master
Change-Id: I282841fdb09c9c2773ca6e5518fdda863f9f8ef5
Reviewed-on: http://review.typo3.org/42854
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>