[BUGFIX] Extbase: Make transOrigDiffSourceField available in DataMap
The transOrigDiffSourceField needs to be available in the DataMap so
that the persistence layer can properly initialize that field when
inserting new objects into the database. This field is often set to
NOT NULL in the schema which leads to SQL errors when it's not included
in the records, for example with MySQL in strict mode.

Resolves: #69085
Releases: master
Change-Id: I021ac695ab495d3d7257dfae27bd596d49551ccf
Reviewed-on: http://review.typo3.org/42654
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>