[BUGFIX] Handling of uploaded files not within open_basedir
When the "upload_tmp_dir" folder is not within the "open_basedir"
an Exception will get thrown when uploading files in the fileadmin.
This patch silences open_basedir() warnings occuring in the file_exists()
check and adds detection of uploaded files as the handling of these with
move_uploaded_file() is not influenced by the open_basedir setting.

Resolves: #69356
Releases: master, 6.2
Change-Id: I2e6b14f207cfd88de04e522e8bcc1bfce72cb0e1
Reviewed-on: https://review.typo3.org/42888
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Markus Sommer <markussom@posteo.de>
Tested-by: Markus Sommer <markussom@posteo.de>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>