[BUGFIX] Form NumberFields ignore typed values in wizard
"Spinner fields" in EXT:form are an old remake of the HTML5
number in ExtJS. Directly typing a value into that spinner
field is not stored - only by clicking the accordant buttons
to "spinup" or "spindown" the value.

A listener for the blur event handles this task and now allows
to store typed numbers as well. The event handler utilizes the
blur event to update the value - thus, the disabled onBlur
handling has to be disabled as well.
Using the change event did not work out due to ExtJS constraints.

Resolves: #69024
Releases: master, 6.2
Change-Id: I13264019c3edbfb8aa7051a4ae2e3ce91dd5c7f6
Reviewed-on: http://review.typo3.org/42895
Tested-by: Bjoern Jacob <bjoern.jacob@tritum.de>
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-on: http://review.typo3.org/42911