[BUGFIX] Abstract XSD generator can handle more than two namespace parts
The getTagNameForClass method in the abstract XSD generator class
could not handle class names with more than two namespace parts.

This caused e.g. TYPO3\CMS\Fluid\ViewHelpers\Be\Buttons\CshViewHelper to
not be transformed to "be.buttons.csh", but just "be.", which in turn
caused duplicate id errors in the resulting XSD document.

Make the code aware of more than two namespace parts.

Releases: master, 6.2
Resolves: #69463
Change-Id: I9a6a28b470679e7487f0d83a7ce9f267a1f3ec7f
Reviewed-on: http://review.typo3.org/42971
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>