[BUGFIX] Streamline Unit Tests for GeneralUtility::sanitizeLocalUrl
Currently the Unit Tests for the function just work by accident. As
HTTP_HOST an absolute path is set which is used to get some other values
to compare the given strings. To add further tests we should streamline
the tests and set/add expectations for host and script_name.

Resolves: #69475
Releases: master, 6.2
Change-Id: If139da7bd22c1ebf2c538cd3412fb0abf56dc668
Reviewed-on: http://review.typo3.org/42982
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>