[TASK] Install Tool: Show database errors on initial import
Database error that occur during the initial install will not be
silently suppressed anymore. The user will be shown the statement and
the error reported by the database.

To halt the installation at this step the state of the initial data
import needs to be tracked and persisted. After all statements have been
processed successfully a key will be added to the system registry
marking the import as done.

Resolves: #51930
Releases: master
Change-Id: I21f9f85524344f2744ac70a54610775ae86c8ea2
Reviewed-on: http://review.typo3.org/43028
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>