[BUGFIX] PageLayoutView: Honor page permissions in getTable_tt_content()
The call to readPageAccess() expects a SQL fragment to check the page
permissions but the raw integer value for the access rights was passed.
The resulting invalid part of the SQL statement (for example `AND 31`)
is ignored by MySQL but results in the permissions not being checked.

Translate the page permissions to the expected SQL fragment for
readPageAccess().

Resolves: #69660
Releases: master
Change-Id: I51a92948dd754410f128683892a9312d97067ea3
Reviewed-on: http://review.typo3.org/43186
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>