[FEATURE] Introduce internal uniqueId generator
This patch add the function StringUtility::getUniqueId() which gets
an unique id (with using more entropy for Windows users) and removes
the included dot to be able to use the generated id in HTML as the dot
is an invalid character in some tag attributes e.g. id.

Resolves: #69730
Releases: master
Change-Id: Ic8a83df0a8b02f6778f0f5588d237379d909b59e
Reviewed-on: http://review.typo3.org/43244
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>