[BUGFIX] Do not create folder in wrong directory
The testing bootstrap creates several folder so that
the tests can run properly. However this is also done
if the tests are executed from a wrong directory
which makes the tests fail. The folders are created anyways.

To fix this we check for an essential file right after the
directory path is determined.

Resolves: #69767
Releases: master
Change-Id: Ic85a12bd1f682b4e45e39031a1b3bda69f0afcdc
Reviewed-on: http://review.typo3.org/43283
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>