[BUGFIX] Warning when clearing all caches from within install tool
When using the "clearAllCache" button a warning message gets shown
(or eventually just logged) when the APC file cache is used.

Reason is that the method "OpcodeCacheUtility::clearAllActive()"
will not allow to clear whole directories. Per definition the used
method "apc_delete_file()" allows only to delete single file-cache
entries from the APC cache when called with a string argument.

So for this to work the method "apc_delete_file()" got changed to
use the "ApcIterator" class using a regular expression matching
the given directory.

Change-Id: I7148fb3c176e05518901b032eccddfa2dd448c4a
Resolves: #69773
Releases: 6.2
Reviewed-on: https://review.typo3.org/43289
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>